x86/mm: Fix build and boot with PG_SIZE != PTE_SIZE
When PG_SHIFT != PTE_SHIFT, PAGE_SIZE/PAGE_SHIFT/PAGE_MASK are
intentionally left undefined to flush out code that conflates the two
granularities. This broke x86 assembly files and exposed several mm
bugs.

Assembly fixes (PAGE_* -> PTE_*):
- arch/x86/entry/calling.h: PTI_USER_PGTABLE_BIT uses PTE_SHIFT (bit
  12 for 4K hardware page boundary between kernel/user PGD halves)
- arch/x86/entry/entry_32.S: PTI_SWITCH_MASK same reasoning
- All other .S files: PAGE_SIZE/PAGE_SHIFT/PAGE_MASK replaced with
  PTE_SIZE/PTE_SHIFT/PTE_MASK since x86 assembly always refers to the
  hardware page size (4K), which is PTE_SIZE in the new scheme

Buddy allocator fix (mm/internal.h, mm/page_alloc.c):
- find_buddy_page_pfn(), buddy_merge_likely(), __free_one_page(): When
  buddy_pfn < pfn, the unsigned subtraction wraps to a huge value.
  Dividing by PTES_PER_PAGE before adding to a struct page pointer
  breaks two's complement wraparound (e.g., 0xFFFFFFFFFFFFFC00/16 =
  0x0FFFFFFFFFFFFFFC0 instead of the correct -64). Fixed by casting
  to (long) before division.

pgdat_end_pfn fix (include/linux/mmzone.h):
- node_spanned_pages is stored in PG_SIZE units (via PTES_TO_PAGES()
  in zone_spanned_pages_in_node()), but pgdat_end_pfn() added it
  directly to node_start_pfn (PTE_SIZE units). Added * PTES_PER_PAGE
  to match zone_end_pfn() which already has this scaling.

Tested: boots to login prompt with PG_SIZE=PTE_SIZE=4K. PG_SIZE>PTE_SIZE
reaches NixOS Stage 1 but hits pre-existing page mapping bugs.

+++ mm/internal.h
@@ -778,7 +778,7 @@ static inline struct page *find_buddy_page_pfn(struct page *page,
 	unsigned long __buddy_pfn = __find_buddy_pfn(pfn, order);
 	struct page *buddy;
 
-	buddy = page + (__buddy_pfn - pfn)/PTES_PER_PAGE;
+	buddy = page + (long)(__buddy_pfn - pfn) / PTES_PER_PAGE;
 	if (buddy_pfn)
 		*buddy_pfn = __buddy_pfn;
